+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.2
 
-- Added `@subtransform!` macro that performs a mutating transform on a subset of rows of a DataFrame.
+- Added `@where!` macro that performs a mutating transform on a subset of rows of a DataFrame.
 
 ## v0.1.2